# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ActivityAction": {
        "enum": [
          "created",
          "edited",
          "deleted"
        ],
        "title": "ActivityAction",
        "type": "string"
      },
      "ActivityOut": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActivityAction"
          },
          "amount_paise": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount Paise"
          },
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "txn_id": {
            "title": "Txn Id",
            "type": "string"
          },
          "txn_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TxnType"
              },
              {
                "type": "null"
              }
            ]
          },
          "undoable": {
            "title": "Undoable",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "txn_id",
          "action",
          "at",
          "title",
          "amount_paise",
          "txn_type",
          "undoable"
        ],
        "title": "ActivityOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PinnedBoard": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PinnedUse"
            },
            "title": "Items",
            "type": "array"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/PinnedSuggestion"
            },
            "title": "Suggestions",
            "type": "array"
          }
        },
        "required": [
          "items",
          "suggestions"
        ],
        "title": "PinnedBoard",
        "type": "object"
      },
      "PinnedIn": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "amount_paise": {
            "exclusiveMinimum": 0.0,
            "title": "Amount Paise",
            "type": "integer"
          },
          "category_id": {
            "title": "Category Id",
            "type": "string"
          },
          "sort_order": {
            "default": 0,
            "title": "Sort Order",
            "type": "integer"
          },
          "title": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "amount_paise",
          "category_id",
          "account_id"
        ],
        "title": "PinnedIn",
        "type": "object"
      },
      "PinnedOut": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "title": "Account Id",
            "type": "string"
          },
          "amount_paise": {
            "exclusiveMinimum": 0.0,
            "title": "Amount Paise",
            "type": "integer"
          },
          "category_id": {
            "title": "Category Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "sort_order": {
            "title": "Sort Order",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "amount_paise",
          "category_id",
          "account_id",
          "sort_order",
          "created_at",
          "updated_at"
        ],
        "title": "PinnedOut",
        "type": "object"
      },
...
      "SealOut": {
        "additionalProperties": false,
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "sealed_at": {
            "format": "date-time",
            "title": "Sealed At",
            "type": "string"
          }
        },
        "required": [
          "date",
          "sealed_at"
        ],
        "title": "SealOut",
        "type": "object"
      },
...
      "StampIn": {
        "additionalProperties": false,
        "properties": {
          "at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          }
        },
        "title": "StampIn",
        "type": "object"
      },
...
      "TxnType": {
        "enum": [
          "expense",
          "income",
          "transfer"
        ],
        "title": "TxnType",
        "type": "string"
      },
      "UndoResult": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/ActivityAction"
          },
          "txn": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TxnOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "action",
          "txn"
        ],
        "title": "UndoResult",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "BudgetBox",
    "version": "1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/activities": {
      "get": {
        "description": "The audit trail, newest first. `txn_id` narrows it to one entry's rewrites \u2014\nwhat the editor's 'rewritten twice \u00b7 last on 12 Jul' whisper reads.",
        "operationId": "recent_activities_v1_activities_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "txn_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Txn Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActivityOut"
                  },
                  "title": "Response Recent Activities V1 Activities Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Recent Activities",
        "tags": [
          "ledger"
        ]
      }
    },
    "/v1/activities/{activity_id}/undo": {
      "post": {
        "operationId": "undo_activity_v1_activities__activity_id__undo_post",
        "parameters": [
          {
            "in": "path",
            "name": "activity_id",
            "required": true,
            "schema": {
              "title": "Activity Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UndoResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Undo Activity",
        "tags": [
          "ledger"
        ]
      }
    },
...
    "/v1/pinned": {
      "get": {
        "operationId": "list_pinned_v1_pinned_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PinnedOut"
                  },
                  "title": "Response List Pinned V1 Pinned Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Pinned",
        "tags": [
          "ledger"
        ]
      }
    },
    "/v1/pinned/board": {
      "get": {
        "description": "The pinned manager in one call: pins with their stamp counts, plus repeats\nworth pinning that aren't pinned yet.",
        "operationId": "pinned_board_v1_pinned_board_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinnedBoard"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Pinned Board",
        "tags": [
          "ledger"
        ]
      }
    },
    "/v1/pinned/{pinned_id}": {
      "delete": {
        "operationId": "delete_pinned_v1_pinned__pinned_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "pinned_id",
            "required": true,
            "schema": {
              "title": "Pinned Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Pinned",
        "tags": [
          "ledger"
        ]
      },
      "put": {
        "operationId": "upsert_pinned_v1_pinned__pinned_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "pinned_id",
            "required": true,
            "schema": {
              "title": "Pinned Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinnedIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinnedOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Pinned",
        "tags": [
          "ledger"
        ]
      }
    },
    "/v1/pinned/{pinned_id}/stamp": {
      "post": {
        "operationId": "stamp_pinned_v1_pinned__pinned_id__stamp_post",
        "parameters": [
          {
            "in": "path",
            "name": "pinned_id",
            "required": true,
            "schema": {
              "title": "Pinned Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/StampIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Data"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxnOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stamp Pinned",
        "tags": [
          "ledger"
        ]
      }
    },
...
    "/v1/seals": {
      "get": {
        "operationId": "seals_between_v1_seals_get",
        "parameters": [
          {
            "in": "query",
            "name": "from_day",
            "required": true,
            "schema": {
              "format": "date",
              "title": "From Day",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to_day",
            "required": true,
            "schema": {
              "format": "date",
              "title": "To Day",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SealOut"
                  },
                  "title": "Response Seals Between V1 Seals Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Seals Between",
        "tags": [
          "ledger"
        ]
      }
    },
    "/v1/seals/{day}": {
      "delete": {
        "description": "Reopen a closed page. Idempotent \u2014 an unsealed day unseals fine.",
        "operationId": "unseal_day_v1_seals__day__delete",
        "parameters": [
          {
            "in": "path",
            "name": "day",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Day",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Unseal Day",
        "tags": [
          "ledger"
        ]
      },
      "put": {
        "operationId": "seal_day_v1_seals__day__put",
        "parameters": [
          {
            "in": "path",
            "name": "day",
            "required": true,
            "schema": {
              "format": "date",
              "title": "Day",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SealOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Seal Day",
        "tags": [
          "ledger"
        ]
      }
    },
...
    "/v1/txns": {
      "get": {
        "operationId": "list_txns_v1_txns_get",
        "parameters": [
          {
            "in": "query",
            "name": "from_day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From Day"
            }
          },
          {
            "in": "query",
            "name": "to_day",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To Day"
            }
          },
          {
            "in": "query",
            "name": "category_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Account Id"
            }
          },
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TxnType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxnPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Txns",
        "tags": [
          "ledger"
        ]
      }
    },
    "/v1/txns/recent-amounts": {
      "get": {
        "description": "The amounts this category is usually written for \u2014 the add sheet's whisper.",
        "operationId": "recent_amounts_v1_txns_recent_amounts_get",
        "parameters": [
          {
            "in": "query",
            "name": "category_id",
            "required": true,
            "schema": {
              "title": "Category Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 3,
              "maximum": 10,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecentAmount"
                  },
                  "title": "Response Recent Amounts V1 Txns Recent Amounts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Recent Amounts",
        "tags": [
          "ledger"
        ]
      }
    },
    "/v1/txns/suggest": {
      "get": {
        "operationId": "suggest_titles_v1_txns_suggest_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 6,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TitleSuggestion"
                  },
                  "title": "Response Suggest Titles V1 Txns Suggest Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Suggest Titles",
        "tags": [
          "ledger"
        ]
      }
    },
    "/v1/txns/{txn_id}": {
      "delete": {
        "operationId": "delete_txn_v1_txns__txn_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "txn_id",
            "required": true,
            "schema": {
              "title": "Txn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Txn",
        "tags": [
          "ledger"
        ]
      },
      "get": {
        "operationId": "get_txn_v1_txns__txn_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "txn_id",
            "required": true,
            "schema": {
              "title": "Txn Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxnOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Txn",
        "tags": [
          "ledger"
        ]
      },
      "patch": {
        "operationId": "patch_txn_v1_txns__txn_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "txn_id",
            "required": true,
            "schema": {
              "title": "Txn Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TxnPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxnOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Patch Txn",
        "tags": [
          "ledger"
        ]
      },
      "put": {
        "operationId": "upsert_txn_v1_txns__txn_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "txn_id",
            "required": true,
            "schema": {
              "title": "Txn Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TxnIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxnOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Txn",
        "tags": [
          "ledger"
        ]
      }
    },
...
  }
}